Mixin -> ForestTraversal.h
