+++ gtsam/base/tests/testPriorityScheduler.cpp
@@ -8,17 +8,13 @@
 
 /**
  * @file testPriorityScheduler.cpp
- * @brief Unit tests for PriorityScheduler with forest TaskMixin helpers.
+ * @brief Unit tests for PriorityScheduler scheduling behavior.
  * @author Frank Dellaert
  * @date May, 2025
  */
 
 #include <CppUnitLite/TestHarness.h>
 #include <gtsam/base/PriorityScheduler.h>
-
-#include <iostream>
-#include <memory>
-#include <vector>
 
 using namespace gtsam;
 
@@ -33,151 +29,11 @@ TEST(PriorityScheduler, ReturnValueComposition) {
   auto right = scheduler.schedule(0, [] { return size_t{4}; }).share();
 
   // Parent task waits on the child futures and combines their values.
-  auto parent =
-      scheduler.schedule(10, [left, right]() mutable { return left.get() + right.get(); });
+  auto parent = scheduler.schedule(
+      10, [left, right]() mutable { return left.get() + right.get(); });
 
   const size_t expectedSum = 7;
   EXPECT_LONGS_EQUAL(expectedSum, parent.get());
-}
-
-/* ************************************************************************* */
-// Forest mixin tests
-/* ************************************************************************* */
-
-/// Common TreeNode definition (using shared_ptr).
-struct TreeNode {
-  int value;
-  explicit TreeNode(int v) : value(v) {}
-};
-
-namespace {
-/// Build a small tree for tests.
-template <typename NodeT>
-std::shared_ptr<NodeT> createSimpleTree() {
-  auto n1 = std::make_shared<NodeT>(1);
-  auto n2 = std::make_shared<NodeT>(2);
-  auto n3 = std::make_shared<NodeT>(3);
-  auto n4 = std::make_shared<NodeT>(4);
-  auto n5 = std::make_shared<NodeT>(5);
-  auto n6 = std::make_shared<NodeT>(6);
-  auto n7 = std::make_shared<NodeT>(7);
-  n1->children() = {n2, n3, n4};
-  n2->children() = {n5, n6};
-  n6->children() = {n7};
-  return n1;
-}
-
-}  // namespace
-
-/* ************************************************************************* */
-// Test 1: Bottom-up traversal
-/* ************************************************************************* */
-/// TreeNode specialization for bottom-up test.
-struct TestNode1 : public TreeNode {
-  using TreeNode::TreeNode;
-  std::vector<std::shared_ptr<TestNode1>> children_;
-  size_t subtreeCount{0};
-
-  /// Return child list for traversal.
-  std::vector<std::shared_ptr<TestNode1>>& children() { return children_; }
-
-  /// Bottom-up payload for tests.
-  void bottomUpValue() {
-    // Each node stores 1 plus the sum of its children's results.
-    size_t sum = 1;
-    for (const auto& child : children()) {
-      sum += child->subtreeCount;
-    }
-    subtreeCount = sum;
-  }
-};
-
-/// Forest wrapper for bottom-up traversal.
-struct TestForest1 : public TaskMixin<TestForest1, TestNode1> {
-  std::vector<std::shared_ptr<TestNode1>> roots_;
-
-  const std::vector<std::shared_ptr<TestNode1>>& roots() const {
-    return roots_;
-  }
-
-  /// Count nodes using a bottom-up traversal.
-  size_t countNodes() {
-    runBottomUp(&TestNode1::bottomUpValue);
-    size_t total = 0;
-    for (const auto& root : roots_) {
-      total += root->subtreeCount;
-    }
-    return total;
-  }
-};
-
-TEST(PriorityScheduler, TaskMixinBottomUpCounting) {
-  std::shared_ptr<TestNode1> root = createSimpleTree<TestNode1>();
-  TestForest1 forest;
-  forest.roots_.push_back(root);
-  size_t rootValue = forest.countNodes();
-
-  const size_t expectedTotalNodes = 7;
-  EXPECT_LONGS_EQUAL(expectedTotalNodes, rootValue);
-}
-
-/* ************************************************************************* */
-// Test 2: Top-down traversal
-/* ************************************************************************* */
-/// TreeNode specialization for top-down test.
-struct TestNode2 : public TreeNode {
-  using TreeNode::TreeNode;
-  std::vector<std::shared_ptr<TestNode2>> children_;
-  bool visited{false};
-  size_t subtreeCount{0};
-
-  /// Return child list for traversal.
-  std::vector<std::shared_ptr<TestNode2>>& children() { return children_; }
-
-  /// Top-down payload for tests.
-  void topDownTouch() { visited = true; }
-
-  /// Bottom-up payload for tests.
-  void bottomUpVisitedCount() {
-    // Each node stores 1 if visited plus the sum of its children.
-    size_t sum = visited ? 1 : 0;
-    for (const auto& child : children()) {
-      sum += child->subtreeCount;
-    }
-    subtreeCount = sum;
-  }
-};
-
-/// Forest wrapper for top-down traversal.
-struct TestForest2 : public TaskMixin<TestForest2, TestNode2> {
-  std::vector<std::shared_ptr<TestNode2>> roots_;
-
-  const std::vector<std::shared_ptr<TestNode2>>& roots() const {
-    return roots_;
-  }
-
-  /// Touch all nodes using a top-down traversal.
-  void touchAll() { runTopDown(&TestNode2::topDownTouch); }
-
-  /// Count visited nodes using a bottom-up traversal.
-  size_t countVisited() {
-    runBottomUp(&TestNode2::bottomUpVisitedCount);
-    size_t total = 0;
-    for (const auto& root : roots_) {
-      total += root->subtreeCount;
-    }
-    return total;
-  }
-};
-
-TEST(PriorityScheduler, TaskMixinTopDown) {
-  std::shared_ptr<TestNode2> root = createSimpleTree<TestNode2>();
-  TestForest2 forest;
-  forest.roots_.push_back(root);
-  forest.touchAll();
-
-  const size_t expectedTotalNodes = 7;
-  EXPECT_LONGS_EQUAL(expectedTotalNodes, forest.countVisited());
 }
 
 /* ************************************************************************* */
